+++ services/web/src/app/[locale]/(private)/layout.tsx
@@ -1,8 +1,18 @@
-import { PropsWithChildren } from 'react';
-import { SessionProvider } from 'next-auth/react';
+import React, { PropsWithChildren } from 'react';
+// import Header from '@/components/common/header'
+import Header from '@/components/header';
+import { withAuth } from '@/components/auth/with-auth';
 
-export default async function Layout({
+async function Layout({
   children,
 }: Readonly<PropsWithChildren>) {
-  return <>{children}</>;
+  return (
+    <div className="flex flex-1 flex-col bg-neutral-100">
+      <Header />
+      {children}
+      {/*<Footer/>*/}
+    </div>
+  );
 }
+
+export default withAuth(Layout);