+++ services/web/src/auth.ts
@@ -1,4 +1,4 @@
 import NextAuth from "next-auth"
-import {authConfig} from '@/config/auth.config';
+import { authConfig } from '@/config/auth.config';
 
 export const {handlers, signIn, signOut, auth} = NextAuth({...authConfig});